requirements.txt
selenium==3.12.0
python-crontab==2.3.4
pymongo==3.6.1
python-dotenv==0.8.2